feat: remove monthly summaries table
This commit removes the monthly summaries table and related
relations from the database schema.

The monthly summaries table was deemed unnecessary for the
current functionality of the application. The data it contained
can be calculated on demand from the transactions table.
Removing the table simplifies the database schema and reduces
the amount of data that needs to be stored.

fix(drizzle.config.ts): remove tablesFilter to generate all tables
fix(README.md): update todo list